improve sensor SOLID and fix bugs, implement base functionality of datalake-builder and analytics

+++ sensor/src/main/java/WeatherResponse.java
@@ -54,7 +54,7 @@ public class WeatherResponse {
         public Wind wind;
         public Rain rain;
         public Clouds clouds;
-        public int dt;
+        public long dt;
         public Sys sys;
         public int timezone;
         public int id;
